refactor draw / state queueing
