+++ src/RageUtil/Graphics/Display/CommandBatcher.cpp
@@ -1,28 +1,30 @@
 #include "CommandBatcher.h"
 #include <cassert>
 
-void Display::CommandBatcher::InsertCommand(const Command &command)
-{
-    assert(command.type != CommandType::Draw);
 
-    m_CommandBuffer.push_back(command);
-	m_RenderStateCount += command.type == CommandType::RenderStateChanged;
+void
+Display::CommandBatcher::InsertRenderStateCommand(
+  RenderState renderState)
+{
+	m_RenderStateBuffer.push_back(renderState);
 }
 
-void Display::CommandBatcher::InsertDrawCommand(DrawMode drawMode, MatrixState matrixState,
+void
+Display::CommandBatcher::InsertDrawCommand(DrawMode drawMode,
+										   MatrixState matrixState,
                                                 const RageSpriteVertex *vertexData, int vertexCount)
 {
     assert(drawMode != DrawMode::Invalid);
 
-    Command command = {};
-    command.type = CommandType::Draw;
-    command.draw.drawMode = drawMode;
-    command.draw.matrices = matrixState;
-    command.draw.vertex = vertexData;
-    command.draw.vertexCount = vertexCount;
+    DrawCommand command = {.drawMode = drawMode, .matrices = matrixState};
 
-	assert(m_RenderStateCount >= 1 && "Rendering information must be set before drawing");
-	command.draw.renderStateIndex = m_RenderStateCount - 1;
+	// TODO: adjust stuff based on draw mode
+    command.vertex = vertexData;
+    command.vertexCount = vertexCount;
+
+	assert(m_RenderStateBuffer.size() >= 1 &&
+		   "Rendering information must be set before drawing");
+	command.renderStateIndex = m_RenderStateBuffer.size() - 1;
 
     m_CommandBuffer.push_back(command);
 }
@@ -30,5 +32,5 @@ void Display::CommandBatcher::InsertDrawCommand(DrawMode drawMode, MatrixState m
 void Display::CommandBatcher::Clear()
 {
     m_CommandBuffer.clear();
-	m_RenderStateCount = 0;
+	m_RenderStateBuffer.clear();
 }
